WIP: introduce cert into dc-agent container
This commit add volume of a cert from a secret, mount the cert into
the container and trust it before start the api service.

This commit also adds some env veriables into the container.
However, the SSL failure still exists after trusting those certs.
Verified the cert was stored in the /etc/ssl/certs/ca-certificates.crt
which is the cert bundle in a normal stx env.